Djanog-4-Boiler-Plate-master-master/requirement.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==4.0.5
django-cleanup==6.0.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
jwt==1.3.1
passlib==1.7.4
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
